docker/requirements.txt
Django==5.0.7
daphne==4.1.2
django-eventstream==5.2.3
djangorestframework==3.15.1